feat: clean up
