+++ lib/cmds/organization_cmds/taxonomy-transform.ts
@@ -9,13 +9,8 @@ import { cursorPaginate } from '../../utils/cursor-pagninate'
 import { ensureDir, getPath, readFileP, writeFileP } from '../../utils/fs'
 import { getHeadersFromOption } from '../../utils/headers'
 import { success, log } from '../../utils/log'
-import {
-  ConceptProps,
-  ConceptSchemeProps,
-  CreateConceptProps,
-  CreateConceptSchemeProps
-} from 'contentful-management'
 import * as Papa from 'papaparse'
+import { Taxonomy } from './taxonomy'
 
 module.exports.command = 'taxonomy-transform'
 
@@ -76,426 +71,7 @@ interface Params {
   silent?: boolean
 }
 
-const defaultLocale = 'en-US'
-
-export class Taxonomy {
-  private concepts: {
-    existing: Array<Concept>
-    added: Array<Concept>
-    updated: Array<Concept>
-  }
-
-  private conceptSchemes: {
-    existing: Array<ConceptScheme>
-    added: Array<ConceptScheme>
-    updated: Array<ConceptScheme>
-  }
-
-  constructor() {
-    this.concepts = {
-      existing: [],
-      added: [],
-      updated: []
-    }
-
-    this.conceptSchemes = {
-      existing: [],
-      added: [],
-      updated: []
-    }
-  }
-
-  toJson() {
-    const updatedConcepts = this.concepts.updated.map(
-      concept => concept.toJson().id
-    )
-    const updatedConceptsSchemes = this.conceptSchemes.updated.map(
-      conceptScheme => conceptScheme.toJson().id
-    )
-
-    return {
-      concepts: [
-        ...this.concepts.existing
-          .filter(concept => !updatedConcepts.includes(concept.toJson().id))
-          .map(concept => concept.toJson()),
-        ...this.concepts.added.map(concept => concept.toJson()),
-        ...this.concepts.updated.map(concept => concept.toJson())
-      ],
-      conceptSchemes: [
-        ...this.conceptSchemes.existing
-          .filter(
-            conceptScheme =>
-              !updatedConceptsSchemes.includes(conceptScheme.toJson().id)
-          )
-          .map(conceptScheme => conceptScheme.toJson()),
-        ...this.conceptSchemes.added.map(conceptScheme =>
-          conceptScheme.toJson()
-        ),
-        ...this.conceptSchemes.updated.map(conceptScheme =>
-          conceptScheme.toJson()
-        )
-      ]
-    }
-  }
-
-  setExistingConcepts(concepts: Array<ConceptProps>) {
-    this.concepts.existing = concepts.map(
-      concept => new Concept(concept.sys.id, concept)
-    )
-  }
-
-  setExistingConceptSchemes(conceptSchemes: Array<ConceptSchemeProps>) {
-    this.conceptSchemes.existing = conceptSchemes.map(
-      conceptScheme => new ConceptScheme(conceptScheme.sys.id, conceptScheme)
-    )
-  }
-
-  addConcept(
-    id: string,
-    init: Partial<Omit<CreateConceptProps, 'id'>> & {
-      prefLabel: CreateConceptProps['prefLabel']
-    }
-  ) {
-    if (this.concepts.existing.find(concept => concept.toJson().id === id)) {
-      return null
-    }
-
-    if (this.concepts.added.find(concept => concept.toJson().id === id)) {
-      return null
-    }
-
-    const concept = new Concept(id, init)
-
-    this.concepts.added.push(concept)
-
-    return concept
-  }
-
-  getConcept(id: string) {
-    const existingConcept = this.concepts.existing.find(
-      concept => concept.toJson().id === id
-    )
-
-    if (!existingConcept) {
-      return null
-    }
-
-    this.concepts.updated.push(existingConcept)
-
-    return existingConcept
-  }
-
-  addConceptScheme(
-    id: string,
-    init: Partial<Omit<CreateConceptSchemeProps, 'id'>> & {
-      prefLabel: CreateConceptSchemeProps['prefLabel']
-    }
-  ) {
-    if (
-      this.conceptSchemes.existing.find(
-        conceptScheme => conceptScheme.toJson().id === id
-      )
-    ) {
-      return null
-    }
-
-    const conceptScheme = new ConceptScheme(id, init)
-
-    this.conceptSchemes.added.push(conceptScheme)
-
-    return conceptScheme
-  }
-
-  getConceptScheme(id: string) {
-    const existingConceptScheme = this.conceptSchemes.existing.find(
-      conceptScheme => conceptScheme.toJson().id === id
-    )
-
-    if (!existingConceptScheme) {
-      return null
-    }
-
-    this.conceptSchemes.updated.push(existingConceptScheme)
-
-    return existingConceptScheme
-  }
-}
-
-class ConceptScheme {
-  private model: CreateConceptSchemeProps & { id: string }
-
-  public constructor(
-    id: string,
-    init: Partial<Omit<CreateConceptSchemeProps, 'id'>> & {
-      prefLabel: CreateConceptSchemeProps['prefLabel']
-    }
-  ) {
-    this.model = { id, ...init }
-  }
-
-  toJson() {
-    return structuredClone(this.model)
-  }
-
-  setUri(uri: string | null) {
-    this.model.uri = uri
-    return this
-  }
-
-  setDefinition(definition: string | null) {
-    if (!this.model.definition) {
-      this.model.definition = {}
-    }
-    this.model.definition[defaultLocale] = definition
-    return this
-  }
-
-  setPrefLabel(prefLabel: string) {
-    if (!this.model.prefLabel) {
-      this.model.prefLabel = {}
-    }
-    this.model.prefLabel[defaultLocale] = prefLabel
-    return this
-  }
-
-  addTopConcept(conceptId: string) {
-    if (!this.model.topConcepts) {
-      this.model.topConcepts = []
-    }
-    this.model.topConcepts.push({
-      sys: {
-        id: conceptId,
-        type: 'Link',
-        linkType: 'TaxonomyConcept'
-      }
-    })
-
-    this.model.totalConcepts = (this.model.totalConcepts || 0) + 1
-
-    return this
-  }
-
-  removeTopConcept(conceptId: string) {
-    if (this.model.topConcepts) {
-      this.model.topConcepts = this.model.topConcepts.filter(
-        concept => concept.sys.id !== conceptId
-      )
-    }
-    if (this.model.concepts) {
-      this.model.concepts = this.model.concepts.filter(
-        concept => concept.sys.id !== conceptId
-      )
-    }
-
-    this.model.totalConcepts = this.model.totalConcepts
-      ? this.model.totalConcepts - 1
-      : 0
-
-    return this
-  }
-
-  addConcept(conceptId: string) {
-    if (!this.model.concepts) {
-      this.model.concepts = []
-    }
-    this.model.concepts.push({
-      sys: {
-        id: conceptId,
-        type: 'Link',
-        linkType: 'TaxonomyConcept'
-      }
-    })
-
-    this.model.totalConcepts = (this.model.totalConcepts || 0) + 1
-
-    return this
-  }
-
-  removeConcept(conceptId: string) {
-    if (this.model.concepts) {
-      if (!this.model.concepts.find(concept => concept.sys.id === conceptId)) {
-        throw new Error(
-          `Concept with id ${conceptId} does not exist in Concept Scheme`
-        )
-      }
-
-      this.model.concepts = this.model.concepts.filter(
-        concept => concept.sys.id !== conceptId
-      )
-
-      this.model.totalConcepts = this.model.totalConcepts
-        ? this.model.totalConcepts - 1
-        : 0
-    }
-
-    return this
-  }
-}
-
-class Concept {
-  private model: CreateConceptProps & { id: string }
-
-  public constructor(
-    id: string,
-    init: Partial<Omit<CreateConceptProps, 'id'>> & {
-      prefLabel: CreateConceptProps['prefLabel']
-    }
-  ) {
-    this.model = { id, ...init }
-  }
-
-  toJson() {
-    return structuredClone(this.model)
-  }
-
-  setUri(uri: string | null) {
-    this.model.uri = uri
-    return this
-  }
-
-  setDefinition(definition: string | null) {
-    if (!this.model.definition) {
-      this.model.definition = {}
-    }
-    this.model.definition[defaultLocale] = definition
-    return this
-  }
-
-  addAltLabel(altLabel: string) {
-    if (!this.model.altLabels) {
-      this.model.altLabels = { [defaultLocale]: [] }
-    }
-    this.model.altLabels[defaultLocale].push(altLabel)
-    return this
-  }
-
-  removeAltLabel(altLabel: string) {
-    if (this.model.altLabels) {
-      this.model.altLabels[defaultLocale] = this.model.altLabels[
-        defaultLocale
-      ].filter(label => label !== altLabel)
-    }
-    return this
-  }
-
-  addHiddenLabel(hiddenLabel: string) {
-    if (!this.model.hiddenLabels) {
-      this.model.hiddenLabels = { [defaultLocale]: [] }
-    }
-    this.model.hiddenLabels[defaultLocale].push(hiddenLabel)
-    return this
-  }
-
-  removeHiddenLabel(hiddenLabel: string) {
-    if (this.model.hiddenLabels) {
-      this.model.hiddenLabels[defaultLocale] = this.model.hiddenLabels[
-        defaultLocale
-      ].filter(label => label !== hiddenLabel)
-    }
-    return this
-  }
-
-  setEditorialNote(editorialNote: string | null) {
-    if (!this.model.editorialNote) {
-      this.model.editorialNote = {}
-    }
-    this.model.editorialNote[defaultLocale] = editorialNote
-    return this
-  }
-
-  setHistoryNote(historyNote: string | null) {
-    if (!this.model.historyNote) {
-      this.model.historyNote = {}
-    }
-    this.model.historyNote[defaultLocale] = historyNote
-    return this
-  }
-
-  setExample(example: string | null) {
-    if (!this.model.example) {
-      this.model.example = {}
-    }
-    this.model.example[defaultLocale] = example
-    return this
-  }
-
-  setNote(note: string | null) {
-    if (!this.model.note) {
-      this.model.note = {}
-    }
-    this.model.note[defaultLocale] = note
-    return this
-  }
-
-  setScopeNote(scopeNote: string | null) {
-    if (!this.model.scopeNote) {
-      this.model.scopeNote = {}
-    }
-    this.model.scopeNote[defaultLocale] = scopeNote
-    return this
-  }
-
-  addNotation(notation: string) {
-    if (!this.model.notations) {
-      this.model.notations = []
-    }
-    this.model.notations.push(notation)
-    return this
-  }
-
-  removeNotation(notation: string) {
-    if (this.model.notations) {
-      this.model.notations = this.model.notations.filter(n => n !== notation)
-    }
-    return this
-  }
-
-  addBroader(conceptId: string) {
-    if (!this.model.broader) {
-      this.model.broader = []
-    }
-    this.model.broader.push({
-      sys: {
-        id: conceptId,
-        type: 'Link',
-        linkType: 'TaxonomyConcept'
-      }
-    })
-    return this
-  }
-
-  removeBroader(conceptId: string) {
-    if (this.model.broader) {
-      this.model.broader = this.model.broader.filter(
-        concept => concept.sys.id !== conceptId
-      )
-    }
-    return this
-  }
-
-  addRelated(conceptId: string) {
-    if (!this.model.related) {
-      this.model.related = []
-    }
-    this.model.related.push({
-      sys: {
-        id: conceptId,
-        type: 'Link',
-        linkType: 'TaxonomyConcept'
-      }
-    })
-    return this
-  }
-
-  removeRelated(conceptId: string) {
-    if (this.model.related) {
-      this.model.related = this.model.related.filter(
-        concept => concept.sys.id !== conceptId
-      )
-    }
-    return this
-  }
-}
+export const defaultLocale = 'en-US'
 
 interface TransformContext {
   csv: {
